requirements.txt
webexteamssdk>=1.6.1
python-dotenv>=1.0.0
openai>=1.0.0
flask>=3.0.0
requests>=2.31.0
